+++ src/main/java/faang/school/accountservice/controller/RequestController.java
@@ -21,6 +21,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 @RequestMapping("api/v1/requests")
 public class RequestController {
+
     private final RequestService requestService;
 
     @PostMapping()
