+++ src/main/java/faang/school/accountservice/mapper/RequestMapper.java
@@ -9,6 +9,7 @@ import org.mapstruct.ReportingPolicy;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface RequestMapper {
+
     Request toEntity(CreateRequestDto createRequestDto);
 
     @Mapping(target = "id", source = "token")
