+++ src/main/java/faang/school/accountservice/service/RequestService.java
@@ -7,9 +7,8 @@ import java.util.Map;
 import java.util.UUID;
 
 public interface RequestService {
+
     void createRequest(CreateRequestDto createRequestDto);
-
     void updateRequestStatusByToken(UUID requestToken, RequestStatus newRequestStatus);
-
     void updateRequestBodyByToken(UUID requestToken, Map<String, Object> newBody);
 }
